Add Frozen JSON parser and emitter library
- Introduced `frozen.h`, a comprehensive JSON parser and emitter for C/C++.
- Implemented key functionalities including `json_scanf()`, `json_printf()`, `json_setf()`, and file handling functions like `json_fread()` and `json_fprintf()`.
- Added support for base64 encoding/decoding and a callback-based API for low-level parsing.
- Documented the API and usage examples in `frozen_json.md`, detailing the features and providing code snippets for common operations.
- Ensured compliance with ISO C and C++ standards, maintaining a small footprint and no external dependencies.

+++ src/graphics/imgpngfilters.c
@@ -199,7 +199,7 @@ BOOL applyPNGFilters(UBYTE *decompressedData, ULONG decompressedSize,
             break;
 
         default:
-            sprintf(logMessage, "Unknown PNG filter type: %d in row %lu", filterType, row);
+            sprintf(logMessage, "Unknown PNG filter type: %d in row %lu", (int)filterType, (unsigned long)row);
             fileLoggerAddDebugEntry(logMessage);
             return FALSE;
         }
